Add --ignore option
Allow to skip missing files. For example, network role for now
don't need .travis/runcoveralls.sh, but as sync-template.sh
just simply copies files listed on its white list, we made a special
branch in template repo just for network role that does not
contain coverage reporting stuff and extend sync-template.sh
about skipping missing files and directories functionality
so it does not fail if desired files or directories are not
found in the file system.