add script to list open PRs and issues in all LSR repos
The script will list all of the open PRs and issues in all of
the LSR repos, sorted from oldest create date to newest.
The script relies on the `hub` and `jq` commands.
Use the env. var. `LISTTYPE` to control the output
`LISTTYPE=both` will print a list of PRs then a list of issues.
`LISTTYPE=prs` will print a list of PRs
`LISTTYPE=issues` will print a list of issues