RHELPLAN-59755 - Collection - script - not include symlinks in the collection converted format
In the commit 6442eafcd69c91808ef6e11daf09d5c53d541f8d, symlinks in the
tests dir are converted to the real files and directories including
the module_utils, which is not preferable. With this commit, such sym-
links targeting the outside of the tests are removed. Only symlinks
which points inside the tests are untouched. For instance,
lrwxrwxrwx. <<snip>> network/tests/playbooks/files -> ../files/