Fixing the dependencies value type in galaxy.yml
Note:
Collections that this collection requires to be installed for it to be
usable. The key of the dict is the collection label 'namespace.name'.
The value is a version range
L(specifiers,https://python-semanticversion.readthedocs.io/en/latest/#requirement-specification).
Multiple version range specifiers can be set and are separated by ','

dependencies: {}