support for ansible 2.11 and other versions
Add support for ansible 2.11 and other versions.  Get rid of
the hardcoded `2.9` version.  The ansible-test ignore files
are now generated for multiple versions of ansible, depending
on whatever the role has.
the .ansible-lint for the collection is now generated by
combining the ones from the roles.
Combine the meta/requirements.yml from each role into the
galaxy.yml dependencies