add support for --check-fqcn
This allows us to check for usage of plugins without FQCN.  In
collections, you must use FQCN for non-module plugins such as
filters, tests, etc. that are not built-in.  This applies to plugins
that are defined by the role - Ansible will not automatically look
them up in the collection if the role and the plugin are in
the same collection.  Now you can use --check-fqcn=plugins.

Example:

Found 9 errors
ERROR: not FQCN: vpn_ipaddr at roles/vpn/tasks/mesh_conf.yml:14

The role in the collection should be using namespace.name.vpn_ipaddr