use git commit subjects for changelog if no version
If we are using the commit hash to build the collection, there will
not be a version and a corresponding changelog.  In that case, use
the git commit titles and list them as Bug Fixes.
* This will inevitably be wrong - there will be New Features
and Other Changes listed erroneously - but there is no way to
know
* When there is an actual release with a version and a changelog,
some of the changelog information will be duplicated - the new
release will list some of the old changelog items again.

I think this is a minor problem to have as long as releases using
commit hashes are rare, and the changelog isn't really used in
Galaxy anyway.

Add support for `--use-commit-hash-role ROLENAME`.  For example,
if you want to do a collection release using the tag for every
role *except* sshd and network, you can do this:
```
./release_collection.py --use-commit-hash-role sshd \
  --use-commit-hash-role network
```