adding a new role - use ref: null; added ad_integration and podman
Added ad_integration and podman roles.

When you want to add a new role to the collection, add it in alphabetical order
to collection_release.yml with a `ref: null`.  The next time you use
release_collection.py to build the collection, it will update the ref with the
correct tag.  Also, the changelog will say
```

...
- ROLENAME - New Role
...
```