Add lsr_fingerprint.py
`lsr_fingerprint` scans files in the `templates` dir under `./role["lsrrolename"]`,
if the file contains a string `role["reponame"]:role["rolename"]`,
replaces it with `system_role:role["lsrrolename"]`. Note: `role` is an entry
of `roles` that is defined as an array of dictionaries in `lsr_fingerprint.py`.

For example, in metrics, `performancecopilot:ansible-pcp` is replaced with
`system_role:metrics`; in sshd, `willshersystems:ansible-sshd` is with
`system_role:sshd`.