Fetch new includes from Fedora distgit
Before Packit builds the SRPM, it needs to fetch files that are missing
from the auto-maintenance repo because they are maintained in Fedora
dist-git and not here. This includes the spec file itself and any files
that it %includes. When new include files (or any other file) are added
to dist-git, they need to be added to the Packit configuration as well.
Do that for the two newly added include snippets:
vendoring-prep.inc and vendoring-build.inc.

https://src.fedoraproject.org/rpms/linux-system-roles/pull-request/203
Fixes #267