ruamel now requires insertion of new key then deletion
The old code would delete the old key using `pop` which would also
retrieve the value of the old key, then insert the new key with the
value at the old index.  This causes ruamel 0.17.23 and later to
crash with an "index out of range" error.  Instead, retrieve the value,
then insert the value with the new key at the old index, then finally
delete the old key.