manage_issues.py with ability create multiple issues and epics
This version of manage_issues.py allows creation of multiple issues and
epics at the same time, and links the created issues to the created
epic.  It also allows creation of issues from github issues and PRs,
similar to manage-issues.sh.  There is still plenty of missing
functionality of manage-issues.sh.