write errors inline in log, parse them out
I think it is a lot of trouble to have to manage another output file
in all of our test environments.  It is much simpler to just write
the error JSON blob inline in the regular output, with tokens marking
the start and end of the blob.  Then parsing it out is easy.